requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy

# tests/linter
coverage
flake8
pytest
yapf

# API
gcsfs
s3fs

# utilities
six>=1.14
memoized-property
termcolor

streamlit
cv2
PIL
tensorflow.keras
matplotlib.image
